requirements.txt
Django==1.10.4
Markdown==2.6.7
django-filter==1.0.1
djangorestframework==3.5.3
wsgiref==0.1.2
